# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Genesis Trade Suite API
  version: 1.0.0
  description: >
    Production REST API for Genesis Trade Suite, providing Shopify-native
    trade accounts, real-time credit controls, double-entry AR ledger receipts,
    multi-document payment allocations, and eight-bucket aging analysis.
servers:
  - url: https://api.genesistradesuite.com/v1
    description: Production API
  - url: http://127.0.0.1:3001/v1
    description: Local development API
security:
  - ShopifySessionToken: []
paths:
...
  /payments:
    post:
      summary: Record payment or receipt
      description: Posts double-entry journal (DR Bank/Cash, CR Accounts Receivable) with visible payment mode.
      operationId: recordPayment
      tags:
        - Payments
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
      responses:
        '201':
          description: Payment posted successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /receipts:
    post:
      summary: Alias for POST /payments
      description: Record an evidence-backed receipt or manual payment.
      operationId: recordReceipt
      tags:
        - Payments
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
      responses:
        '201':
          description: Receipt posted successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
...
components:
  securitySchemes:
    ShopifySessionToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Signed Shopify App Bridge session token HMAC-SHA256 verified.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Unique idempotency key preventing duplicate command execution.
      schema:
        type: string
  schemas:
...
    PaymentRequest:
      type: object
      required:
        - debtorAccountId
        - amount
        - currency
        - paymentMethod
        - effectiveDate
        - reference
        - paymentMode
      properties:
        debtorAccountId:
          type: string
          format: uuid
        amount:
          type: string
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
        paymentMethod:
          type: string
          enum: [bank_transfer, cash, card, shopify_manual_payment, check]
        effectiveDate:
          type: string
          format: date
        reference:
          type: string
        paymentMode:
          type: string
          enum: [external_receipt, shopify_manual, shopify_pos_cash]
          description: Visible payment synchronization mode.
        memo:
          type: string
        autoAllocate:
          type: boolean
          default: false
    PaymentResponse:
      type: object
      required:
        - journalId
        - documentId
        - paymentMode
        - paymentMethod
        - amount
        - currency
        - effectiveDate
        - reference
        - isDuplicate
        - unappliedCredit
        - netBalance
      properties:
        journalId:
          type: string
          format: uuid
        documentId:
          type: string
          format: uuid
        paymentMode:
          type: string
          enum: [external_receipt, shopify_manual, shopify_pos_cash]
        paymentMethod:
          type: string
        amount:
          type: string
        currency:
          type: string
        effectiveDate:
          type: string
          format: date
        reference:
          type: string
        isDuplicate:
          type: boolean
        unappliedCredit:
          type: string
        netBalance:
          type: string
        allocations:
          type: array
          items:
            type: object
            properties:
              allocationId: { type: string, format: uuid }
              debitDocumentId: { type: string, format: uuid }
              amount: { type: string }
...
    ErrorResponse:
      type: object
      required:
        - error
        - code
        - message
      properties:
        error:
          type: string
        code:
          type: string
        message:
          type: string
  responses:
    BadRequest:
      description: Missing required fields or malformed request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication failed or missing valid session token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated actor lacks required permission.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Requested resource was not found for this tenant.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: Concurrency version conflict, duplicate identifier, or idempotency conflict with differing payload.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnprocessableEntity:
      description: Business policy rejection, credit declined, or currency mismatch.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'